backend/requirements.txt
fastapi
uvicorn
minio
psycopg2-binary
sqlalchemy
python-multipart
python-dotenv
pandas
numpy
matplotlib
scikit-learn
mlxtend
statsmodels
watchdog
python-jose[cryptography]
requests